Day4

+++ Adventureland-Village/src/main/java/com/adventurelandVillage/dto/CustomerTicketDTO.java
@@ -20,6 +20,6 @@ public class CustomerTicketDTO {
 
 	private List<Ticket> tickets;
 
-	private Integer totalAmount;
+	private float totalAmount;
 
 }
